Refactor file directories
